requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.1
bandit==1.8.6
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
Django==4.2.23
djangorestframework==3.16.1
drf-spectacular==0.28.0
exceptiongroup==1.3.0
freezegun==1.5.5
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
model-bakery==1.20.5
packaging==25.0
pbr==7.0.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.9
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
